OneDrive/Desktop/IDCC_Application/requirements-minimal.txt
# Minimal requirements for basic functionality
# Use this if full installation fails

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
aiofiles==23.2.1
PyPDF2==3.0.1
python-docx==1.1.0

# Optional - install separately if needed
# numpy>=1.26.0
# pandas==2.1.3
# scikit-learn==1.3.2
# pillow==10.1.0
# opencv-python==4.8.1.78
# pytesseract==0.3.10

